+++ userProfile/index.html
@@ -7,7 +7,7 @@
 </head>
 
 <body>
-    <div id="main"></div>
+    <div id="root"></div>
     <script src="build/index.js"></script>
 </body>
